integration-test: make status report test more robust
An exam might already be running when the integration test is
launched, so it's not robust nor very relevant to assert the run state
in this state.

But because the point of this is to test if status reporting is
working, we check that examomatic has received some status reports and
that the latest one is fresh.